Fix warnings and a little more

+++ src/dos/dos_misc.cpp
@@ -490,7 +490,7 @@ static bool DOS_MultiplexFunctions(void) {
             uint32_t size = 0, extra = 0;
             unsigned char head, last=13;
             uint8_t *text;
-            for (int i=0; i<strPasteBuffer.length(); i++) if (strPasteBuffer[i]==10||strPasteBuffer[i]==13) extra++;
+            for (size_t i=0; i<strPasteBuffer.length(); i++) if (strPasteBuffer[i]==10||strPasteBuffer[i]==13) extra++;
             if (strPasteBuffer.length() && (text = (uint8_t *)malloc(strPasteBuffer.length()+extra))) {
                 while (strPasteBuffer.length()) {
                     head = strPasteBuffer[0];
@@ -528,7 +528,7 @@ static bool DOS_MultiplexFunctions(void) {
             uint32_t size = 0, extra = 0;
             unsigned char head, last=13;
             uint8_t *text;
-            for (int i=0; i<strPasteBuffer.length(); i++) if (strPasteBuffer[i]==10||strPasteBuffer[i]==13) extra++;
+            for (size_t i=0; i<strPasteBuffer.length(); i++) if (strPasteBuffer[i]==10||strPasteBuffer[i]==13) extra++;
             if (strPasteBuffer.length() && (text = (uint8_t *)malloc(strPasteBuffer.length()+extra))) {
                 while (strPasteBuffer.length()) {
                     head = strPasteBuffer[0];
